add response suffix and move bcs types to general types file
