+++ ecosystem/typescript/sdk_v2/src/bcs/deserializer.ts
@@ -3,7 +3,7 @@
 
 /* eslint-disable no-bitwise */
 import { MAX_U32_NUMBER } from "./consts";
-import { Uint128, Uint16, Uint256, Uint32, Uint64, Uint8 } from "./types";
+import { Uint128, Uint16, Uint256, Uint32, Uint64, Uint8 } from "../types";
 
 // The class must implement a static deserialize method.
 interface Deserializable<T> {
