+++ ecosystem/typescript/sdk_v2/src/bcs/index.ts
@@ -1,6 +1,5 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-export * from "./types";
 export * from "./serializer";
 export * from "./deserializer";
