--- ecosystem/typescript/sdk_v2/src/bcs/types.ts
@@ -1,10 +1,0 @@
-// Copyright © Aptos Foundation
-// SPDX-License-Identifier: Apache-2.0
-
-export type Uint8 = number;
-export type Uint16 = number;
-export type Uint32 = number;
-export type Uint64 = bigint;
-export type Uint128 = bigint;
-export type Uint256 = bigint;
-export type AnyNumber = bigint | number;
